requirements.txt
absl-py==0.8.1
astor==0.8.0
certifi==2019.11.28
chardet==3.0.4
contextlib2==0.6.0.post1
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
gast==0.3.2
grpcio==1.16.1
h5py==2.9.0
idna==2.8
imageio==2.6.1
kiwisolver==1.1.0
lxml==4.4.2
Markdown==3.1.1
matplotlib==3.1.2
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
msgpack==0.6.2
msgpack-numpy==0.4.4.3
networkx==2.4
numpy==1.17.4
opencv-python==4.1.2.30
Pillow==6.2.1
pinyin==0.4.0
protobuf==3.11.2
psutil==5.6.7
pycocotools==2.0
pyparsing==2.4.6
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.2
pyzmq==18.1.1
requests==2.22.0
scipy==1.3.2
six==1.13.0
tabulate==0.8.6
tensorboard>=1.13.1
tensorflow>=1.13.1
tensorflow-estimator>=1.13.0
termcolor==1.1.0
tqdm==4.41.1
urllib3==1.25.7
Werkzeug==0.16.0
yacs==0.1.6
